1 Translate the SimpleKeywordTest class to English.
